// ui/src/generated/SearchFilesCursor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DirectoryScanFrame } from "./DirectoryScanFrame";

export type SearchFilesCursor = { project: string, text: string, show_hidden: boolean, directories: Array<DirectoryScanFrame>, };